Image 3.0.0 not showing
The images are now displaying.
Maybe it could be moved to another line to better looking. But I couldn't make it.
Other option could be: 
Move ". First we will store all the updated grading information in a relational database. It will also stream separately to the shipping departments." between the two images and making it be part of the list.